Edit value in tableview

+++ src/main/java/com/example/demo1/Page2.java
@@ -3,21 +3,23 @@ package com.example.demo1;
 import com.example.demo1.sqlOperation.GeneralUtils;
 import com.example.demo1.sqlOperation.MysqlInterface;
 import javafx.application.Application;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableArray;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.Alert;
-import javafx.scene.control.TablePosition;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.stage.Stage;
 
 import java.io.IOException;
-import java.util.Arrays;
 
 public class Page2 extends Application {
 
 
     public TableView tableview;
+    public ChoiceBox choiceBox1;
+    public TextField stock_field;
+    public TextField prix_field;
+    public TextField taille_field;
 
     @Override
     public void start(Stage primaryStage) {
@@ -25,6 +27,9 @@ public class Page2 extends Application {
     }
     @FXML
     protected void initialize(){
+
+        var type_produit = new String[]{"Chaussure","Vêtement","Accesoire"};
+        choiceBox1.setItems(FXCollections.observableArrayList(type_produit));
 
         tableview.getItems().clear();
 
@@ -46,24 +51,43 @@ public class Page2 extends Application {
     }
 
     @FXML
-    private void tableviewSelect(){
+    private void AddStock() throws IOException{
 
-        var row_val = new GeneralUtils().getRowSelected2StrArray(tableview);
+        var type_produit = choiceBox1.getSelectionModel().getSelectedItem();
+        var prix = prix_field.getText();
+        var taille = taille_field.getText();
+        var stock = stock_field.getText();
+        boolean allIsNum = new GeneralUtils().isInt(taille) && new GeneralUtils().isInt(stock) && new GeneralUtils().isDouble(prix);
 
-        var row_eval = new GeneralUtils().isInt(row_val[0]) ? Integer.parseInt(row_val[0]) : row_val[0];
-        System.out.println(" /row0 "+ row_eval);
+        if (type_produit != null && prix!=null && (taille!=null || type_produit == "Accessoir") && stock !=null && allIsNum){
+            if(type_produit == "Accessoir"){
+                taille = "null";
+            }
+            new MysqlInterface().WriteData("Insert into produit(Categorie,Taille, Prix, Stock) values ('"
+                    + type_produit + "',"+taille+","+prix+","+stock+");");
 
+            Alert a = new Alert(Alert.AlertType.INFORMATION,"Stock has been added !");
+            a.show();
+
+            switchToSecondePage();
+        }
+        else {
+            Alert a = new Alert(Alert.AlertType.WARNING,"Champs mal remplis");
+            a.show();
+        }
     }
     @FXML
     private void SupprimerStock(){
 
         var row_val = new GeneralUtils().getRowSelected2StrArray(tableview);
-        new MysqlInterface().WriteData("Delete from produit where Id ="+row_val[0]+";");
+        if(row_val != null) {
+            new MysqlInterface().WriteData("Delete from produit where Id =" + row_val[0] + ";");
 
-        Alert a = new Alert(Alert.AlertType.INFORMATION,"Row has been deleted");
-        a.show();
+            Alert a = new Alert(Alert.AlertType.INFORMATION,"Row has been deleted");
+            a.show();
 
-        initialize();
+            initialize();
+        }
     }
     @FXML
     private void SupprimeItem() throws IOException{
